backend/pyproject.toml
[tool.poetry]
name = "moodify-backend"
version = "0.1.0"
description = "Backend for music recommendation app"
authors = ["Gurshan Sidhu"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.108.0"
uvicorn = {extras = ["standard"], version = "^0.25.0"}
openai = "^1.6.0"
spotipy = "^2.22.1"
python-dotenv = "^1.0.0"
pydantic = "^2.5.0"
python-multipart = "^0.0.6"
requests = "^2.32.5"

[tool.poetry.group.dev.dependencies]
black = "^23.12.0"
isort = "^5.13.0"
pytest = "^7.4.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py311']

[tool.isort]
profile = "black"
line-length = 88
